Fixed reading of plugin config
The plugin config file contains the configuration of the box
files as we create it as a service for the users. It is not
meant to be edited by users in the first place. Therefore the
location in /etc is somewhat misleading. This commit moves
the basic box configuration to be part of the plugin source.
As also the box image descriptions are handled in this git
changes to the box setup and eventually to the plugin config
are now going together. As a possitive side effect this also
fixes the installation of the plugin from pip which did
not properly handle the custom install routine for the config
file when it is hosted in /etc.