Added --no-snapshot option
Run box with snapshot mode switched off. This causes the
box disk file to be modified by the build process and allows
to keep a persistent package cache as part of the box.
The option can be used to increase the build performance
due to data stored in the box which doesn't have to be
reloaded from the network. On the contrary this option
invalidates the immutable box attribute and should be
used with care. On update of the box all data stored
will be wiped. To prevent this combine the option with
the --no-update-check option.