Enable the arm architecture in boxbuild
Add new options --aarch64, --cpu and --machine which allows
to select and configure cross arch builds in particular for
the arm architecture in this commit. A box build for aarch64
now exists for our users and can be used with the proposed
changes. To allow a cross arch build for arm on x86 the
following command can be used

    kiwi-ng system boxbuild \
        --box universal --cpu cortex-a57 --machine virt -- ...