Store box files in the users HOME
The boxbuild command is intented to be called as normal user
and not as root. The former location /var/tmp/kiwi/boxes pins
the permissions of the downloaded box files to the user who
did it first. This can lead to unexpected side effects when
different users on the same system uses boxbuild.
This Fixes #28