Fixed creation of the build command passed to kiwi
The provided data in 'kiwi_build_command_args' from the
doc string of the kiwi boxbuild command is passed on as argument
to the kiwi build command inside of the box.

However the argument list is not fully compatible because for
the boxbuild command 'kiwi_build_command_args' is the argument
to the '--' option whereas for the build command any element
of what was provided in 'kiwi_build_command_args' is an option + value
for the build command.

Therefore it was possible to say "-- --description=foo" which
ends up as '--description=foo' to be passed as one option without
value to the build command which of course does not exist.

To fix it, this pull request makes sure that the information given
in the 'kiwi_build_command_args' is parsed using docopt against
the kiwi.tasks.system_build doc string and the result
of that parsing operation is used to construct the kiwi build
command.

This Fixes #33