Use progressbar python module
As alternative to the no longer developed wget module
we are switching to the progressbar module. This requires
the code to fetch the data to be implemented through the
requests module to allow utilizing the progressbar as
it is designed. A nice after effect is that the progressbar
displayed is more sophisticated than the one before