Support sshfs as sharing backend
Implement support for sshfs sharing backend in a way that
host and guest SSH ports are redirected such that sshfs
can be used with KVM user networking

The sshfs operation is performed from the guest using the
host forwared SSH port. To make this call a non-interactive
process it's required that the host accepts the box ssh
pub key as authorized_keys. The implementation intentionally
does not change the host's authorized_keys file which is the
reason why this setup will be added into the documentation.

Without adding the box pubkey into the authorized_keys
file of the user on the host which calls boxbuild the process
will stop asking for a passphrase or depending on the hosts
sshd configuration it could also fail if e.g password auth
is disabled.

This Fixes #44