Fixed initrd lookup
Due to a change in kiwi with regards to the compression
algorithm used for initrd's in the distribution, the lookup
of the initrd file in boxbuild broke. kiwi uses by default
--xz but this is no longer correct for all distros.
Thus kiwi does not longer force xz for compressing initrds
created by dracut. For details see: OSInside/kiwi#1987
This commit fixes the box plugin to work with that kiwi
change.