Add the ability to set the SSH forwarding port on the host
It is not currently possible to run multiple Kiwi instances using box
build at the same time as they will all share the same SSH port. This
patch adds the command line argument `--ssh-port`, which allows the user
to specify an SSH port to use for the connection to the guest. The
argument takes a single positive integer value.

Signed-off-by: Isaac True <isaac.true@canonical.com>