kiwi_boxed_plugin: utils: fetch_files: improve progressbar import
The progressbar2 module is installed on Debian-based systems as
"progressbar2", and can be imported as such, but if progressbar2 has been
installed from pip the module is called "progressbar". Adjust the import of
this module to first try importing "progressbar2", and if it fails, fall back
to "progressbar". This stops the tool from trying to import the old
progressbar module rather than the correct progressbar2 module.

Signed-off-by: Isaac True <isaac.true@canonical.com>