Add cascading config finder
Boxed plugin can only get a static hard-coded configuration, delivered as a resource inside
a Python package. It cannot be overridden without explicit change of a package-delivered data.
This is also makes it impossible to programmatically integrate it with Kiwi as a library. Under
these circumstances, a cascading mechanism is required, which is searching for the configuration
file in the following order:

* $KIWI_BOXED_PLUGIN_CFG full path, the name can be anything
* Current directory, from where a software was called
* $HOME/.config/kiwi/kiwi_boxed_plugin.yml
* /etc/kiwi_boxed_plugin.yml
* Resource (default config, coming with the package)