Make sure to run with exported PATH setting
The kiwi run inside of the box runs from an environment
that has no PATH setup. This can lead to issues when
Path.which() is used to find program files in the box
and also as base for the search inside of the new root
tree. If no PATH is set kiwi uses os.defpath which does
not take enough paths (e.g /usr/sbin) into account.
This commit makes sure we run kiwi with a proper PATH
setup.