Update box versions
Update box versions to match with the repositories
they built for. In addition I learned that download.opensuse.org
uses mirror servers which only updates the content if the file
name changes. The version information of our box files is also
part of the file names and that's the only part of the file
which changes. Thus a simple rebuild of a box file due to a
change in the image description does not update the image binary
on the download mirror. We need to find a proper way to handle
this situation better. For now I just updated all box versions
manually.