Fix box update issue
The box images are built in OBS and fetched from the download server.
The download server uses mirrors to manage the data. The mirror update
is based on the filename of the data. The filename of the box files
only changed when we updated the name or the version which is not
necessarily the case when we fix bugs in the plugin code or when
package updates in the distribution happens. Thus the mirror data
was often old even though a fix was made. To prevent this issue I
reactivated the build counter in the OBS projects building the box
files, such that every rebuild of the image gets a new BuildXX
number in the result files. This of course introduces a problem
with the plugin configuration which needs to reference static
file names. To solve this problem this commit moved to the static
link names produced by OBS.