Add container support to boxbuild
In normal mode boxbuild, builds the image in a KVM virtual
machine. With the new --container option, boxbuild can now
also build the image in a podman controlled container. The
boxes as we have so far received an additional type to also
build as containers. In container mode boxbuild pulls the
respective container from the remote registry to the local
registry and starts a container instance. There are the
following runtime constraints:

* As kiwi requires loop devices and can call other operations
  which requires root privileges, podman is started through sudo.
  As podman runs daemonless, the calling user must have the
  privileges to perform the needed kiwi actions. Non
  privileged builds are therefore not possible in container
  mode with kiwi.

* The following linux capabilities are added to the container
  - AUDIT_WRITE
  - AUDIT_CONTROL
  - CAP_MKNOD
  - CAP_SYS_ADMIN
  For details on this capabilities please refer to
  https://man7.org/linux/man-pages/man7/capabilities.7.html

* The host device nodes (/dev) are shared with the container
  instance. This is required to work with loop devices
  inside of the container. Device isolation is therefore
  not possible in container mode with kiwi.